requirements.txt
fastapi[standard]
uvicorn
numpy
torch
torchvision
streamlit
streamlit-webrtc
facenet-pytorch
opencv-python-headless
efficientnet-pytorch
pillow
pandas
requests
